test(decomposition): split mcp server and hardening monolith suites
