+++ tests/mcp_server/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from fastapi.testclient import TestClient
+
+from analyst_toolkit.mcp_server.server import app
+
+
+@pytest.fixture
+def client() -> TestClient:
+    return TestClient(app)
